Extract hash password function and use it in migrations files
