+++ src/services/auth/auth.logic.setup.ts
@@ -1,8 +1,9 @@
 import * as authLogicFactories from "./auth.logic";
 import { saveQuery } from "../../utils/common/common.query";
-import { checkIsEmailTaken } from "../../utils/auth/auth.query";
+import { checkIsEmailTaken, hashPassword } from "../../utils/auth/auth.utils";
 
 export const registerLogic = authLogicFactories.registerLogicFactory(
   checkIsEmailTaken,
+  hashPassword,
   saveQuery,
 );
