--- src/utils/auth/auth.query.interface.ts
@@ -1,1 +1,0 @@
-export type checkIsEmailTakenFunc = (mail: string) => Promise<void>;
