+++ src/utils/auth/auth.utils.interface.ts
@@ -1,0 +1,3 @@
+export type checkIsEmailTakenFunc = (mail: string) => Promise<void>;
+
+export type hashPasswordFunc = (password: string) => Promise<string>;
