+++ src/utils/auth/auth.utils.ts
@@ -1,7 +1,8 @@
 import { getRepository } from "typeorm";
 import { User } from "../../entities/user/user.model";
 import * as Err from "../errors/errors";
-import { checkIsEmailTakenFunc } from "./auth.query.interface";
+import { checkIsEmailTakenFunc } from "./auth.utils.interface";
+import * as argon2 from "argon2";
 
 export const checkIsEmailTaken: checkIsEmailTakenFunc = (
   mail: string,
@@ -13,3 +14,6 @@ export const checkIsEmailTaken: checkIsEmailTakenFunc = (
         throw Err.MailAlreadyTaken(mail);
       }
     });
+
+export const hashPassword = (password: string): Promise<string> =>
+  argon2.hash(password);
